refactor create Consumables
