+++ lib/constants/utilitis/utilitis.dart
@@ -102,58 +102,55 @@ class Utilitis {
     required Function() onReject,
   }) =>
       showDialog(
-          context: context,
-          barrierColor: const Color.fromARGB(20, 0, 0, 0),
-          builder: (context) => Dialog(
-                backgroundColor: Colors.white,
-                child: SizedBox(
-                  // decoration: BoxDecoration(
-                  // border: Border.all(color: AppColor.kTextfieldBorder, width: 1.5),
-                  // borderRadius: BorderRadius.circular(30)),
-                  height: 350,
-                  width: MediaQuery.of(context).size.width / 10 * 60,
+        context: context,
+        barrierColor: const Color.fromARGB(20, 0, 0, 0),
+        builder: (context) => Dialog(
+          backgroundColor: Colors.white,
+          child: SizedBox(
+            // decoration: BoxDecoration(
+            // border: Border.all(color: AppColor.kTextfieldBorder, width: 1.5),
+            // borderRadius: BorderRadius.circular(30)),
+            height: 350,
+            width: MediaQuery.of(context).size.width / 10 * 60,
+            child: Column(
+              mainAxisAlignment: MainAxisAlignment.center,
+              children: [
+                Padding(
+                  padding: const EdgeInsets.all(8.0),
                   child: Column(
-                    mainAxisAlignment: MainAxisAlignment.center,
                     children: [
-                      Padding(
-                        padding: const EdgeInsets.all(8.0),
-                        child: Column(
-                          children: [
-                            Text(
-                              message,
-                              style: Theme.of(context).textTheme.titleLarge,
-                              textAlign: TextAlign.center,
+                      Text(
+                        message,
+                        style: Theme.of(context).textTheme.titleLarge,
+                        textAlign: TextAlign.center,
+                      ),
+                      Row(
+                        mainAxisAlignment: MainAxisAlignment.center,
+                        children: [
+                          Padding(
+                            padding: const EdgeInsets.symmetric(horizontal: 40, vertical: 40.0),
+                            child: SymmetricButton(
+                              text: 'Ja',
+                              onPressed: onAccept,
+                              padding: const EdgeInsets.symmetric(horizontal: 12, vertical: 4),
                             ),
-                            Row(
-                              mainAxisAlignment: MainAxisAlignment.center,
-                              children: [
-                                Padding(
-                                  padding:
-                                      const EdgeInsets.symmetric(horizontal: 40, vertical: 40.0),
-                                  child: SymmetricButton(
-                                    text: 'Ja',
-                                    onPressed: onAccept,
-                                    padding:
-                                        const EdgeInsets.symmetric(horizontal: 12, vertical: 4),
-                                  ),
-                                ),
-                                Padding(
-                                  padding:
-                                      const EdgeInsets.symmetric(horizontal: 40, vertical: 40.0),
-                                  child: SymmetricButton(
-                                    text: 'Nein',
-                                    onPressed: onReject,
-                                    padding:
-                                        const EdgeInsets.symmetric(horizontal: 12, vertical: 4),
-                                  ),
-                                ),
-                              ],
+                          ),
+                          Padding(
+                            padding: const EdgeInsets.symmetric(horizontal: 40, vertical: 40.0),
+                            child: SymmetricButton(
+                              text: 'Nein',
+                              onPressed: onReject,
+                              padding: const EdgeInsets.symmetric(horizontal: 12, vertical: 4),
                             ),
-                          ],
-                        ),
+                          ),
+                        ],
                       ),
                     ],
                   ),
                 ),
-              ));
+              ],
+            ),
+          ),
+        ),
+      );
 }
