+++ lib/models/consumable_models/consumable_vm/consumable_vm.freezed.dart
@@ -21,9 +21,9 @@ ConsumableVM _$ConsumableVMFromJson(Map<String, dynamic> json) {
 /// @nodoc
 mixin _$ConsumableVM {
   int get amount => throw _privateConstructorUsedError;
-  int get id => throw _privateConstructorUsedError;
+  int? get id => throw _privateConstructorUsedError;
   String get name => throw _privateConstructorUsedError;
-  int get price => throw _privateConstructorUsedError;
+  double get price => throw _privateConstructorUsedError;
   Unit get unit => throw _privateConstructorUsedError;
 
   Map<String, dynamic> toJson() => throw _privateConstructorUsedError;
@@ -38,7 +38,7 @@ abstract class $ConsumableVMCopyWith<$Res> {
           ConsumableVM value, $Res Function(ConsumableVM) then) =
       _$ConsumableVMCopyWithImpl<$Res, ConsumableVM>;
   @useResult
-  $Res call({int amount, int id, String name, int price, Unit unit});
+  $Res call({int amount, int? id, String name, double price, Unit unit});
 
   $UnitCopyWith<$Res> get unit;
 }
@@ -57,7 +57,7 @@ class _$ConsumableVMCopyWithImpl<$Res, $Val extends ConsumableVM>
   @override
   $Res call({
     Object? amount = null,
-    Object? id = null,
+    Object? id = freezed,
     Object? name = null,
     Object? price = null,
     Object? unit = null,
@@ -67,10 +67,10 @@ class _$ConsumableVMCopyWithImpl<$Res, $Val extends ConsumableVM>
           ? _value.amount
           : amount // ignore: cast_nullable_to_non_nullable
               as int,
-      id: null == id
+      id: freezed == id
           ? _value.id
           : id // ignore: cast_nullable_to_non_nullable
-              as int,
+              as int?,
       name: null == name
           ? _value.name
           : name // ignore: cast_nullable_to_non_nullable
@@ -78,7 +78,7 @@ class _$ConsumableVMCopyWithImpl<$Res, $Val extends ConsumableVM>
       price: null == price
           ? _value.price
           : price // ignore: cast_nullable_to_non_nullable
-              as int,
+              as double,
       unit: null == unit
           ? _value.unit
           : unit // ignore: cast_nullable_to_non_nullable
@@ -103,7 +103,7 @@ abstract class _$$ConsumableVMImplCopyWith<$Res>
       __$$ConsumableVMImplCopyWithImpl<$Res>;
   @override
   @useResult
-  $Res call({int amount, int id, String name, int price, Unit unit});
+  $Res call({int amount, int? id, String name, double price, Unit unit});
 
   @override
   $UnitCopyWith<$Res> get unit;
@@ -121,7 +121,7 @@ class __$$ConsumableVMImplCopyWithImpl<$Res>
   @override
   $Res call({
     Object? amount = null,
-    Object? id = null,
+    Object? id = freezed,
     Object? name = null,
     Object? price = null,
     Object? unit = null,
@@ -131,10 +131,10 @@ class __$$ConsumableVMImplCopyWithImpl<$Res>
           ? _value.amount
           : amount // ignore: cast_nullable_to_non_nullable
               as int,
-      id: null == id
+      id: freezed == id
           ? _value.id
           : id // ignore: cast_nullable_to_non_nullable
-              as int,
+              as int?,
       name: null == name
           ? _value.name
           : name // ignore: cast_nullable_to_non_nullable
@@ -142,7 +142,7 @@ class __$$ConsumableVMImplCopyWithImpl<$Res>
       price: null == price
           ? _value.price
           : price // ignore: cast_nullable_to_non_nullable
-              as int,
+              as double,
       unit: null == unit
           ? _value.unit
           : unit // ignore: cast_nullable_to_non_nullable
@@ -156,7 +156,7 @@ class __$$ConsumableVMImplCopyWithImpl<$Res>
 class _$ConsumableVMImpl extends _ConsumableVM {
   const _$ConsumableVMImpl(
       {required this.amount,
-      required this.id,
+      this.id,
       required this.name,
       required this.price,
       required this.unit})
@@ -168,11 +168,11 @@ class _$ConsumableVMImpl extends _ConsumableVM {
   @override
   final int amount;
   @override
-  final int id;
+  final int? id;
   @override
   final String name;
   @override
-  final int price;
+  final double price;
   @override
   final Unit unit;
 
@@ -214,9 +214,9 @@ class _$ConsumableVMImpl extends _ConsumableVM {
 abstract class _ConsumableVM extends ConsumableVM {
   const factory _ConsumableVM(
       {required final int amount,
-      required final int id,
+      final int? id,
       required final String name,
-      required final int price,
+      required final double price,
       required final Unit unit}) = _$ConsumableVMImpl;
   const _ConsumableVM._() : super._();
 
@@ -226,11 +226,11 @@ abstract class _ConsumableVM extends ConsumableVM {
   @override
   int get amount;
   @override
-  int get id;
+  int? get id;
   @override
   String get name;
   @override
-  int get price;
+  double get price;
   @override
   Unit get unit;
   @override
