+++ lib/view/consumable_view/consumable_view.dart
@@ -7,18 +7,26 @@ import '../../models/consumable_models/consumable_vm/consumable_vm.dart';
 import '../../models/consumable_models/unit/unit.dart';
 import '../../provider/consumeable_proivder/consumable_provider.dart';
 import '../shared_view_widgets/search_line_header.dart';
-import 'widgets/card_widget.dart';
+import '../users_view/widgets/add_button_widget.dart';
+import 'widgets/creaet_material_widget.dart';
 import 'widgets/consumeabel_row_widget.dart';
 
 class ConsumableBody extends ConsumerStatefulWidget {
-  const ConsumableBody({super.key});
+  final Duration snackbarDuration;
+
+  const ConsumableBody({
+    super.key,
+    this.snackbarDuration = const Duration(seconds: 7),
+  });
 
   @override
   ConsumerState<ConsumableBody> createState() => _ConsumableBodyState();
 }
 
 class _ConsumableBodyState extends ConsumerState<ConsumableBody> {
-  final List<Unit> _units = [Unit(id: 0, name: 'leer')];
+  bool _isSnackbarShowed = false;
+  late final Duration _snackbarDuration;
+  final List<Unit> _units = [];
   List<ConsumableVM> consumableList = [];
   bool isLoading = true;
   bool isCardVisible = false;
@@ -27,6 +35,7 @@ class _ConsumableBodyState extends ConsumerState<ConsumableBody> {
   void initState() {
     super.initState();
     loadUnits();
+    _snackbarDuration = widget.snackbarDuration;
   }
 
   void loadUnits() {
@@ -49,29 +58,14 @@ class _ConsumableBodyState extends ConsumerState<ConsumableBody> {
   }
 
   void _showSnackBar(String message) {
-    final snackBar = SnackBar(content: Text(message));
-    ScaffoldMessenger.of(context).showSnackBar(snackBar);
-  }
-
-  void hideCard() {
-    setState(() {
-      isCardVisible = false;
-    });
-  }
-
-  void deleteService(ConsumableVM row) async {
-    try {
-      await ref.read(consumableProvider.notifier).deleteConsumable(row.id);
-      // TODO: Sollte View Aktualisieren
-      // ignore: unused_result
-      ref.refresh(consumableProvider);
-      // setState(() {
-      //   consumableList = ref.watch(consumableProvider);
-      // });
-      _showSnackBar('Row deleted successfully.');
-    } catch (e) {
-      _showSnackBar('Error when attempting to delete the item: $e');
-    }
+    if (_isSnackbarShowed) return;
+    setState(() => _isSnackbarShowed = true);
+    ScaffoldMessenger.of(context).showSnackBar(
+      SnackBar(content: Text(message), duration: _snackbarDuration),
+    );
+    Future.delayed(_snackbarDuration).then(
+      (_) => setState(() => _isSnackbarShowed = false),
+    );
   }
 
   @override
@@ -82,47 +76,68 @@ class _ConsumableBodyState extends ConsumerState<ConsumableBody> {
     return buildCardContent();
   }
 
-  Widget buildCardContent() => Container(
-        color: Colors.white,
-        child: Padding(
-          padding: const EdgeInsets.symmetric(horizontal: 30, vertical: 30),
-          child: SingleChildScrollView(
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                const SearchLineHeader(title: 'Material Management'),
-                buildHeaderRow(),
-                (consumableList.isEmpty && _units.isEmpty)
-                    ? const Text('No data available')
-                    : ListView.builder(
+  Widget buildCardContent() => Padding(
+        padding: const EdgeInsets.symmetric(horizontal: 30),
+        child: SingleChildScrollView(
+          child: Column(
+            crossAxisAlignment: CrossAxisAlignment.start,
+            children: [
+              const SearchLineHeader(title: 'Material Management'),
+              buildHeaderRow(),
+              consumableList.isEmpty && _units.isEmpty
+                  ? const Text('No data available')
+                  : SizedBox(
+                      width: MediaQuery.of(context).size.width - 50,
+                      height: 5 * 74,
+                      child: ListView.builder(
                         // shrinkWrap: true,
                         itemCount: consumableList.length,
                         itemBuilder: (context, i) {
                           log('consumables-> ${consumableList.length} units-> ${_units.length}');
                           return ConsumebaleDataRow(
                             consumable: consumableList[i],
-                            onDelete: () => deleteService(consumableList[i]),
+                            onDelete: () {
+                              ref
+                                  .read(consumableProvider.notifier)
+                                  .deleteConsumable(consumableList[i].id!)
+                                  .then((e) {
+                                // ignore: unused_result
+                                ref.refresh(consumableProvider);
+                                e
+                                    ? _showSnackBar('Row deleted successfully.')
+                                    : _showSnackBar('Error when attempting to delete the item: $e');
+                              });
+                            },
+                            // => deleteService(consumableList[i]),
                             units: _units,
                           );
                         },
                       ),
-                const SizedBox(height: 40),
-                buildAddButton(),
-                if (isCardVisible)
-                  CardWidget(
-                    onSave: _addRow,
-                    onHideCard: hideCard,
-                  ),
-              ],
-            ),
+                    ),
+              AddButton(
+                onTap: () => setState(() => isCardVisible = !isCardVisible),
+              ),
+              // const SizedBox(height: 40),
+              // buildAddButton(),
+              if (isCardVisible)
+                CreateMaterialCard(
+                  onSave: _addRow,
+                  //  ref.read(consumableProvider).createConsumable(_, int amount, Unit unit, double price);
+                  onHideCard: () => setState(() => isCardVisible = false),
+                  units: _units,
+                  onAccept: () {
+                    // _createConsumable();
+                  },
+                ),
+            ],
           ),
         ),
       );
 
   Widget buildHeaderRow() => Row(
         children: [
-          Expanded(
-            flex: 3,
+          SizedBox(
+            width: MediaQuery.of(context).size.width / 10 * 2.0,
             child: Text(
               'Material',
               style: Theme.of(context).textTheme.titleLarge?.copyWith(
@@ -130,8 +145,8 @@ class _ConsumableBodyState extends ConsumerState<ConsumableBody> {
                   ),
             ),
           ),
-          Expanded(
-            flex: 2,
+          SizedBox(
+            width: MediaQuery.of(context).size.width / 10 * 2.0,
             child: Text(
               'Menge',
               style: Theme.of(context).textTheme.titleLarge?.copyWith(
@@ -139,8 +154,8 @@ class _ConsumableBodyState extends ConsumerState<ConsumableBody> {
                   ),
             ),
           ),
-          Expanded(
-            flex: 2,
+          SizedBox(
+            width: MediaQuery.of(context).size.width / 10 * 2.0,
             child: Text(
               'Einheit',
               style: Theme.of(context).textTheme.titleLarge?.copyWith(
@@ -148,47 +163,96 @@ class _ConsumableBodyState extends ConsumerState<ConsumableBody> {
                   ),
             ),
           ),
-          Expanded(
-            flex: 2,
-            child: Text(
-              'Preis',
-              style: Theme.of(context).textTheme.titleLarge?.copyWith(
-                    fontWeight: FontWeight.bold,
-                  ),
-            ),
-          ),
-          const Expanded(
-            flex: 1,
-            child: SizedBox(),
+          Text(
+            'Preis',
+            style: Theme.of(context).textTheme.titleLarge?.copyWith(
+                  fontWeight: FontWeight.bold,
+                ),
           ),
         ],
       );
 
-  Widget buildAddButton() => Padding(
-        padding: const EdgeInsets.only(bottom: 8),
-        child: Align(
-          alignment: Alignment.topLeft,
-          child: FloatingActionButton(
-            onPressed: () {
-              setState(() {
-                isCardVisible = !isCardVisible;
-              });
-            },
-            backgroundColor: Colors.orange,
-            child: Icon(isCardVisible ? Icons.remove : Icons.add, color: Colors.white),
-          ),
-        ),
-      );
+  // Widget buildAddButton() => Padding(
+  //       padding: const EdgeInsets.only(bottom: 8),
+  //       child: Align(
+  //         alignment: Alignment.topLeft,
+  //         child: FloatingActionButton(
+  //           onPressed: () {
+  //             setState(() {
+  //               isCardVisible = !isCardVisible;
+  //             });
+  //           },
+  //           backgroundColor: Colors.orange,
+  //           child: Icon(isCardVisible ? Icons.remove : Icons.add, color: Colors.white),
+  //         ),
+  //       ),
+  //     );
 
   void _addRow(String materialName, int amount, Unit unit, int price) {
-    setState(() {
-      consumableList.add(ConsumableVM(
-        id: consumableList.isNotEmpty ? consumableList.last.id + 1 : 1,
-        name: materialName,
-        amount: amount,
-        unit: unit,
-        price: price,
-      ));
-    });
+    // setState(() {
+    //   consumableList.add(ConsumableVM(
+    //     id: consumableList.isNotEmpty ? consumableList.last.id + 1 : 1,
+    //     name: materialName,
+    //     amount: amount,
+    //     unit: unit,
+    //     price: price,
+    //   ));
+    // });
   }
+
+  // void _createConsumable() async {
+  //   final String material = _materialController.text.trim();
+  //   final int amount = int.parse(_amountController.text.trim());
+  //   final Unit unit = _selectedUnit!;
+  //   final int price = int.parse(_priceController.text.trim());
+
+  //   if (!_validateInputs()) {
+  //     return;
+  //   }
+
+  //   try {
+  //     setState(() {
+  //       _isLoading = true;
+  //     });
+
+  //     await ref
+  //         .read(consumableProvider.notifier)
+  //         .createService(material, amount, unit, price)
+  //         .then((value) {
+  //       setState(() {
+  //         _priceController.clear();
+  //         _selectedUnit = null;
+  //         _amountController.clear();
+  //         _materialController.clear();
+  //       });
+  //     });
+
+  //     widget.onSave(material, amount, unit, price);
+
+  //     if (Navigator.canPop(context)) {
+  //       Navigator.of(context).pop();
+  //     }
+  //   } catch (e) {
+  //     log('Error on createService: $e');
+  //     showDialog(
+  //       context: context,
+  //       builder: (ctx) => AlertDialog(
+  //         title: const Text('Fehler beim Speichern'),
+  //         content: Text('Ein Fehler ist aufgetreten: $e'),
+  //         actions: <Widget>[
+  //           TextButton(
+  //             child: const Text('Ok'),
+  //             onPressed: () {
+  //               Navigator.of(ctx).pop();
+  //             },
+  //           ),
+  //         ],
+  //       ),
+  //     );
+  //   } finally {
+  //     setState(() {
+  //       _isLoading = false; // Stop loading after the API call completes
+  //     });
+  //   }
+  // }
 }
