+++ lib/view/customers_view/edit_customer.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
 import '../../constants/themes/app_color.dart';
-import '../administration_view/custom_project.dart';
+import '../../models/project_models/customer_projekt_model/custom_project.dart';
 import '../shared_view_widgets/symetric_button_widget.dart';
 
 class AddNewConsumable extends StatefulWidget {
