+++ lib/view/customers_view/project_card.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import '../administration_view/custom_project.dart';
+import '../../models/project_models/customer_projekt_model/custom_project.dart';
 import 'project_details.dart';
 
 // ignore: must_be_immutable
