+++ lib/view/customers_view/project_report.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import '../administration_view/custom_project.dart';
+import '../../models/project_models/customer_projekt_model/custom_project.dart';
 
 class ProjectReportScreen extends StatefulWidget {
   final CustomeProject project;
