+++ lib/view/project_management_view/edit_project.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
 
-import '../administration_view/custom_project.dart';
+import '../../models/project_models/customer_projekt_model/custom_project.dart';
 import '../shared_view_widgets/symetric_button_widget.dart';
 
 class AddNewProject extends StatefulWidget {
