+++ lib/view/project_management_view/project_managment_view.dart
@@ -2,7 +2,7 @@ import 'dart:developer';
 
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import '../administration_view/custom_project.dart';
+import '../../models/project_models/customer_projekt_model/custom_project.dart';
 import '../shared_view_widgets/search_line_header.dart';
 import '../users_view/widgets/add_button_widget.dart';
 import 'customer_card.dart';
